Updated install_nvim to keep sudo alive during isntall.
This helps during the long build process since sudo normally timesout
and then when the install is run as sudo, the script stops waiting for
user input. After some period, the script will exit if credentials are
not entered. This change should keep sudo alive during the entire script
run.